Add files via upload
Screenshots of  Streamlit and the email for alert message if the crowd is large.  
This shows a sample crowd image from the ShanghaiTech dataset that is used as input to the model.
The corresponding ground truth density map prepared from head annotations in the dataset.
The predicted density map generated by the trained model. Brighter areas represent higher crowd density.